requirements.txt
google-cloud-texttospeech
textract
